+++ ext/jq/recipes/jq_recipe.rb
@@ -3,8 +3,8 @@
 require 'mini_portile2'
 
 class JQRecipe < MiniPortile
-  JQ_VERSION = '1.7.1'
-  JQ_SHA256 = '478c9ca129fd2e3443fe27314b455e211e0d8c60bc8ff7df703873deeee580c2'
+  JQ_VERSION = '1.8.1'
+  JQ_SHA256 = '2be64e7129cecb11d5906290eba10af694fb9e3e7f9fc208a311dc33ca837eb0'
 
   def initialize
     super('jq', JQ_VERSION)